+++ packages/terradart_google/lib/firestore.dart
@@ -47,5 +47,9 @@ export 'src/firestore/google_firestore_index.dart'
         FirestoreIndexIndexFieldTextSpec,
         FirestoreIndexIndexFieldTextSpecEntry,
         FirestoreIndexIndexFieldVectorConfig;
+export 'src/firestore/google_firestore_document.dart'
+    show GoogleFirestoreDocument;
 export 'src/firestore/google_firestore_user_creds.dart'
     show GoogleFirestoreUserCreds;
+export 'src/firestore/firestore_fields.dart'
+    show FirestoreFields, FirestoreReference, FirestoreGeoPoint;